First attempt - set up Templating so that it will build for net451
