+++ src/Microsoft.Framework.CodeGeneration.Common/TemplateFoldersUtilities.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+#if DNX451 || DNXCORE50
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -63,3 +64,4 @@ namespace Microsoft.Framework.CodeGeneration
         }
     }
 }
+#endif
