+++ src/Microsoft.Framework.CodeGeneration.Templating/Compilation/ICompilationService.cs
@@ -1,12 +1,13 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Threading.Tasks;
+using System.Collections.Generic;
+using Microsoft.CodeAnalysis;
 
 namespace Microsoft.Framework.CodeGeneration.Templating.Compilation
 {
     public interface ICompilationService
     {
-        CompilationResult Compile(string content);
+        CompilationResult Compile(string content, List<MetadataReference> references);
     }
 }
