+++ src/Microsoft.Framework.CodeGeneration.Templating/Compilation/RoslynCompilationService.cs
@@ -1,43 +1,20 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.Emit;
-using Microsoft.Framework.Runtime;
-using Microsoft.Framework.Runtime.Roslyn;
 
 namespace Microsoft.Framework.CodeGeneration.Templating.Compilation
 {
     public class RoslynCompilationService : ICompilationService
     {
-        private static readonly ConcurrentDictionary<string, AssemblyMetadata> _metadataFileCache =
-            new ConcurrentDictionary<string, AssemblyMetadata>(StringComparer.OrdinalIgnoreCase);
-
-        private readonly ILibraryManager _libraryManager;
-        private readonly IApplicationEnvironment _environment;
-        private readonly IAssemblyLoadContext _loader;
-
-        public RoslynCompilationService(IApplicationEnvironment environment,
-                                        IAssemblyLoadContextAccessor accessor,
-                                        ILibraryManager libraryManager)
-        {
-            _environment = environment;
-            _loader = accessor.GetLoadContext(typeof(RoslynCompilationService).GetTypeInfo().Assembly);
-            _libraryManager = libraryManager;
-        }
-
-        public CompilationResult Compile(string content)
+        public CompilationResult Compile(string content, List<MetadataReference> references)
         {
             var syntaxTrees = new[] { CSharpSyntaxTree.ParseText(content) };
-
-            var references = GetApplicationReferences();
 
             var assemblyName = Path.GetRandomFileName();
 
@@ -46,7 +23,7 @@ namespace Microsoft.Framework.CodeGeneration.Templating.Compilation
                         syntaxTrees: syntaxTrees,
                         references: references);
 
-            var result = CommonUtilities.GetAssemblyFromCompilation(_loader, compilation);
+            var result = CommonUtilities.GetAssemblyFromCompilation(/*_loader, */ compilation);
             if (result.Success)
             {
                 var type = result.Assembly.GetExportedTypes()
@@ -58,83 +35,6 @@ namespace Microsoft.Framework.CodeGeneration.Templating.Compilation
             {
                 return CompilationResult.Failed(content, result.ErrorMessages);
             }
-        }
-
-        //Todo: This is using application references to compile the template,
-        //perhaps that's not right, we should use the dependencies of the caller
-        //who calls the templating API.
-        private List<MetadataReference> GetApplicationReferences()
-        {
-            var references = new List<MetadataReference>();
-
-            // Todo: When the target app references scaffolders as nuget packages rather than project references,
-            // since the target app does not actully use code from scaffolding framework, our dlls
-            // are not in app dependencies closure. However we need them for compiling the
-            // generated razor template. So we are just using the known references for now to make things work.
-            // Note that this model breaks for custom scaffolders because they may be using
-            // some other references which are not available in any of these closures.
-            // As the above comment, the right thing to do here is to use the dependency closure of
-            // the assembly which has the template.
-            var baseProjects = new string[] { _environment.ApplicationName, "Microsoft.Framework.CodeGeneration" };
-
-            foreach (var baseProject in baseProjects)
-            {
-                var export = _libraryManager.GetAllExports(baseProject);
-
-                foreach (var metadataReference in export.MetadataReferences)
-                {
-                    references.Add(ConvertMetadataReference(metadataReference));
-                }
-            }
-
-            return references;
-        }
-
-        private MetadataReference ConvertMetadataReference(IMetadataReference metadataReference)
-        {
-            var roslynReference = metadataReference as IRoslynMetadataReference;
-
-            if (roslynReference != null)
-            {
-                return roslynReference.MetadataReference;
-            }
-
-            var embeddedReference = metadataReference as IMetadataEmbeddedReference;
-
-            if (embeddedReference != null)
-            {
-                return MetadataReference.CreateFromImage(embeddedReference.Contents);
-            }
-
-            var fileMetadataReference = metadataReference as IMetadataFileReference;
-
-            if (fileMetadataReference != null)
-            {
-                return CreateMetadataFileReference(fileMetadataReference.Path);
-            }
-
-            var projectReference = metadataReference as IMetadataProjectReference;
-            if (projectReference != null)
-            {
-                using (var ms = new MemoryStream())
-                {
-                    projectReference.EmitReferenceAssembly(ms);
-
-                    return MetadataReference.CreateFromImage(ms.ToArray());
-                }
-            }
-
-            throw new NotSupportedException();
-        }
-
-        private MetadataReference CreateMetadataFileReference(string path)
-        {
-            var metadata = _metadataFileCache.GetOrAdd(path, _ =>
-            {
-                return AssemblyMetadata.CreateFromStream(File.OpenRead(path));
-            });
-
-            return metadata.GetReference();
         }
     }
 }
