+++ src/Microsoft.Framework.CodeGeneration.Templating/project.json
@@ -3,13 +3,27 @@
     "dependencies": {
         "Microsoft.Framework.CodeGeneration.Common": { "version": "1.0.0-*", "type": "build" },
         "Microsoft.AspNet.Razor": "4.0.0-*",
-        "Microsoft.Framework.Runtime.Roslyn": "1.0.0-*"
+        "Microsoft.CodeAnalysis.CSharp":  "1.0.0-*"
     },
 
     "frameworks": {
-        "dnx451": { },
+        "net451": {
+            "frameworkAssemblies": {
+                "System.Reflection": "4.0.0.0",
+                "System.Runtime": "4.0.10.0",
+                "System.Collections": "4.0.0.0",
+                "System.Text.Encoding": "4.0.0.0",
+                "System.Threading.Tasks": "4.0.0.0"
+            }
+        },
+        "dnx451": {
+            "dependencies": {
+                "Microsoft.Framework.Runtime.Roslyn": "1.0.0-*"
+            }
+        },
         "dnxcore50": {
             "dependencies": {
+                "Microsoft.Framework.Runtime.Roslyn": "1.0.0-*",
                 "System.Reflection.TypeExtensions": "4.0.0-beta-*"
             }
         }
